Move EventLoop out of the header
The previous implementation required users to be aware of the implementation of EventLoop.
The new version exposes submit_async which is independent of the EventLoop class.
